kafka_consumer/requirements.txt
psycopg2-binary==2.9.9
confluent-kafka==2.2.0
kafka-python==2.0.2
